% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submit_job.R
\name{as_new_job}
\alias{as_new_job}
\title{Create new job request from existing}
\usage{
as_new_job(job = NULL, ..., token = get_current_access_token())
}
\arguments{
\item{job}{A character value with a job ID
or an object that contains job data}

\item{...}{Additional properties to set for new job}

\item{token}{An Encore API access token}
}
\value{
A list containing parameters used
to create job
}
\description{
This function helps to create a new job
request that you can pass to `submit_job`.
The request itself is just a named list with
values that would match what you enter on the
Encore website
}
\details{
You can pass in an existing job to copy all the
existing parameters and just change the ones
you want to be different.

Alternatively, you can pass in named parameters
to override existing values or simply set new ones
}
\examples{
\dontrun{
new_job <- as_new_job(get_jobs(limit=1))
new_job$name <- paste(new_job$name, "Rerun")
submit_job(new_job)
}
}
\seealso{
\code{\link{submit_job}}
}
