% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{download_job_output}
\alias{download_job_output}
\title{Download job output}
\usage{
download_job_output(
  job,
  output_name = NULL,
  verbose = TRUE,
  ...,
  token = get_current_access_token()
)
}
\arguments{
\item{job}{A character vector of job IDs (UUIDs)}

\item{output_name}{Optional file name for download. Will
default to the name used on the server}

\item{verbose}{Show file names and progress of downloads if TRUE}

\item{...}{Optional parameters passed through to `httr::write_disk`}

\item{token}{An Encore API access token}
}
\value{
Invisibly returns the list of files created
}
\description{
Download raw results file from Encore server for given jobs
}
\examples{
\dontrun{
# Download file to current directory
job_id <- "123e4567-e89b-12d3-a456-426652340000"
download_job_output(job_id)
# Download results from ALL jobs
download_job_output(get_jobs())
}
}
