% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encore-package.R
\docType{package}
\name{encore-package}
\alias{encore}
\alias{encore-package}
\title{encore: Client to interact with the Encore analysis server}
\description{
Allows you to interact with an Encore analysis server directly from R.
}
\author{
\strong{Maintainer}: Matthew Flickinger \email{mflick@umich.edu}

}
\keyword{internal}
