% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genos.R
\name{get_genotypes}
\alias{get_genotypes}
\title{Get genotype list}
\usage{
get_genotypes(search = NULL, limit = 200, token = get_current_access_token())
}
\arguments{
\item{search}{A character value used to filter
the genotype list}

\item{limit}{Maximum number of records to return}

\item{token}{An Encore API access token}
}
\value{
A data.frame with class "encore_geno_list"
}
\description{
Return a list of all available genotype freezes
}
\examples{
\dontrun{
get_genotypes()
get_genotypes("Freeze 5")
get_genotypes(limit=1)
}
}
