% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobs.R
\name{get_job_details}
\alias{get_job_details}
\title{Get job details}
\usage{
get_job_details(job, token = get_current_access_token())
}
\arguments{
\item{job}{A job ID string (UUID)}

\item{token}{An Encore API access token}
}
\value{
A object of class "encore_job"
}
\description{
Return the settings used to create a job
}
\examples{
\dontrun{
job_id <- "123e4567-e89b-12d3-a456-426652340000"
get_job_details(job_id)
}
}
\seealso{
\code{\link{get_jobs}}
}
