% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_job_id}
\alias{get_job_id}
\title{Get job ID}
\usage{
get_job_id(job)
}
\arguments{
\item{job}{A character value or an object that
contains job data}
}
\value{
A character value with the job ID(s)
}
\description{
A generic function to determine the job ID for
a given object
}
\examples{
\dontrun{
get_job_id(get_jobs())
}
}
