% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qq.R
\name{get_job_qq_data}
\alias{get_job_qq_data}
\title{Get QQ plot data}
\usage{
get_job_qq_data(job, index, token = get_current_access_token())
}
\arguments{
\item{job}{A job ID string (UUID)}

\item{index}{If multiple QQ plots are returned, the
index for the one you want to use}

\item{token}{An Encore API access token}
}
\value{
This will return an object of type "encore_qq_data"
that can be used with the `plot_qq` function.
}
\description{
This function fetches QQ plot data from API Server
}
\details{
This function will return the data used to draw
the p-value QQ plot on the Encore website. Note
that the data has been summarized on the server
and binned in order to reduce the number of points
to be drawn.
}
\examples{
\dontrun{
job_id <- "123e4567-e89b-12d3-a456-426652340000"
data <- get_job_manhattan_data(job_id)
plot_manhattan(data)
}
}
\seealso{
\code{\link{plot_qq}}
}
