% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobs.R
\name{get_jobs}
\alias{get_jobs}
\title{Get job list}
\usage{
get_jobs(search = NULL, limit = 200, token = get_current_access_token())
}
\arguments{
\item{search}{A character value used to search
the name, date, and ID fields for the jobs}

\item{limit}{Maximum number of records to return}

\item{token}{An Encore API access token}
}
\value{
A data.frame with class "encore_job_list"
}
\description{
Return a list of all your current jobs
}
\examples{
\dontrun{
get_jobs()
get_jobs("saige-qt")
get_jobs(limit=1)
}
}
\seealso{
\code{\link{get_job_details}}, \code{\link{download_job_output}},
\code{\link{get_phenotypes}}, \code{\link{get_genotypes}}
}
