% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenos.R
\name{get_phenotype_details}
\alias{get_phenotype_details}
\title{Get phenotype details}
\usage{
get_phenotype_details(pheno, token = get_current_access_token())
}
\arguments{
\item{pheno}{A phenotype ID string (UUID)}

\item{token}{An Encore API access token}
}
\value{
A object of class "encore_phenotype"
}
\description{
Returns metadata for phenotype file
}
\examples{
\dontrun{
pheno_id <- "123e4567-e89b-12d3-a456-426652340000"
get_phenotype_details(pheno_id)
}
}
\seealso{
\code{\link{get_phenotypes}}
}
