% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_phenotype_id.R
\name{get_phenotype_id}
\alias{get_phenotype_id}
\title{Get phenotype ID}
\usage{
get_phenotype_id(x)
}
\arguments{
\item{x}{A character value or an object that
contains phenotype data}
}
\value{
A character value with the phenotype ID(s)
}
\description{
A generic function to determine the phenotype ID for
a given object
}
\examples{
\dontrun{
get_phenotype_id(get_phenotypes())
}
}
