% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenos.R
\name{get_phenotypes}
\alias{get_phenotypes}
\title{Get phenotype list}
\usage{
get_phenotypes(search = NULL, limit = 200, token = get_current_access_token())
}
\arguments{
\item{search}{A character value used to filter
your phenotype list}

\item{limit}{Maximum number of records to return}

\item{token}{An Encore API access token}
}
\value{
A data.frame with class "encore_pheno_list"
}
\description{
Return a list of all your phenotype files
}
\examples{
\dontrun{
get_phenotypes()
get_phenotypes("my favorite file")
get_phenotypes(limit=1)
}
}
