% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state.R
\name{set_access_token}
\alias{set_access_token}
\title{Set API access token}
\usage{
set_access_token(token)
}
\arguments{
\item{token}{A character value containing Encore API access token.
If you are using RStudio you can leave this value blank and a window
will open to ask for your token. The prevents the token from being stored
in your history file to help keep it more secure.}
}
\description{
Set the default Encore API access token
}
\details{
In order to use the Encore API, you must first
log in and generate an API access token on the
website itself. You'll receive a long string of
letters and numbers (a JSON web token). Most
functions will allow you to pass a token as a
parameter, but by using this function, you set
the default access token you so don't have to
pass it as a parameter to every function you call.

It's important to keep your access token secure.
For some advice on ways to store the value without
typing it in the console, see `vignette("secrets", package="httr")`
}
\examples{
\dontrun{
# This will open a prompt if using RStudio to enter your token
set_access_token()

# Note your actual token will be much longer
# (This example token is not valid)
set_access_token("eyJhbGciOi9.eyJuYW1lIjoiRW5.e30XBiuDHBYDxSLT")
}
}
