% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submit_job.R
\name{submit_job}
\alias{submit_job}
\title{Submit new job request}
\usage{
submit_job(x, token = get_current_access_token())
}
\arguments{
\item{x}{A named list of job parameters}

\item{token}{An Encore API access token}
}
\value{
A object of class "encore_job"
}
\description{
Submit new job request
}
\examples{
\dontrun{
new_job <- as_new_job(get_jobs(limit=1))
new_job$name <- paste(new_job$name, "Rerun")
submit_job(new_job)
}
}
\seealso{
\code{\link{as_new_job}}
}
