% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submit_phenotype.R
\name{submit_phenotype}
\alias{submit_phenotype}
\alias{submit_phenotype.default}
\alias{submit_phenotype.data.frame}
\title{Submit new phenotype file}
\usage{
submit_phenotype(x, ..., token = get_current_access_token())

\method{submit_phenotype}{default}(x, ..., token = get_current_access_token())

\method{submit_phenotype}{data.frame}(x, filename = NULL, ..., token = get_current_access_token())
}
\arguments{
\item{x}{A data.frame or a path to a CSV file}

\item{...}{Additional parameters for S3 methods}

\item{token}{An Encore API access token}

\item{filename}{Override the default CSV name}
}
\value{
An "encore_phenotype" object
}
\description{
Upload phenotype data to Encore server
}
\examples{
\dontrun{
sample <- data.frame(
  id = sprintf("NWD\%06d", 1:100),
  value = rnorm(100),
  group = sample(c("X","Y"), 100, replace=TRUE)
)
submit_phenotype(sample)
}
}
