% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manhattan.R
\name{theme_manhattan}
\alias{theme_manhattan}
\title{Manhattan plot theme}
\usage{
theme_manhattan(..., panel.grid.major = ggplot2::element_blank())
}
\arguments{
\item{...}{Values passed along to the ggplot::theme() function}

\item{panel.grid.major}{Styling chromosome center grid marks}
}
\value{
A ggplot2 object
}
\description{
Default ggplot2 theme for Manhattan plot
}
\details{
This uses the black and white ggplot theme and
turns off the grid lines for the chromosome
centers (major breaks)
}
